+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Please add changes to "master". When releasing a version add a new header for th
  - Batch scripts can now access projects and files as Apache Spark DataFrames and run SQL queries
    on them.
  - Workflows and batch scripts can run workflows.
+ - Improved center picker experience.
 
 ### 1.5.8